# static/swagger.yaml
openapi: 3.0.0
info:
  title: API de Verificación de Rostros y Usuarios
  description: API para la verificación de rostros, inicio de sesión y registro de usuarios.
  version: 1.0.1
servers:
  - url: http://localhost:5000
    description: Servidor local
  - url: https://living-improved-ocelot.ngrok-free.app/
    description: Servidor en linea
...
paths:
...
  /api/signup/teacher:
    post:
      summary: Registrar profesor
      description: Registra un nuevo profesor proporcionando los datos necesarios.
      tags:
        - Registro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nombre completo del profesor.
                username:
                  type: string
                  description: Nombre de usuario.
                age:
                  type: integer
                  description: Edad del profesor.
                faculty:
                  type: string
                  description: Facultad a la que pertenece el profesor.
                worknum:
                  type: string
                  description: Número de trabajo.
                password:
                  type: string
                  description: Contraseña del profesor.
                face_img:
                  type: string
                  format: base64
                  description: Imagen de rostro en base64.
                email:
                  type: string
                  format: email
                  description: Correo electrónico del profesor.
              required:
                - name
                - username
                - age
                - faculty
                - worknum
                - password
                - face_img
                - email
      responses:
        '201':
          description: Registro de profesor exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  status_code:
                    type: integer
        '400':
          description: Faltan campos en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
        '409':
          description: Profesor ya registrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
                  duplicate_field:
                    type: string
                    description: Campo que causa la violación de unicidad (username, worknum o email).
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  status_code:
                    type: integer
...
  /api/student/teachers:
      get:
        summary: Retrieve teachers for a student
        description: Retrieve the list of teachers and their classes for a given student.
        tags:
          - Estudiantes
        parameters:
          - in: query
            name: student_id
            required: true
            schema:
              type: string
            description: The ID of the student.
        responses:
          '200':
            description: A list of teachers and their classes.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    data:
                      type: array
                      items:
                        type: object
                        properties:
                          name:
                            type: string
                          email:
                            type: string
                          class_name:
                            type: array
                            items:
                              type: string
                    status_code:
                      type: integer
          '400':
            description: Missing or invalid student_id parameter.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    error:
                      type: string
                    status_code:
                      type: integer
          '500':
            description: Internal server error.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    error:
                      type: string
                    status_code:
                      type: integer
...